+++ tpl/cli.sh
@@ -8,9 +8,9 @@ cat ${BASE_CONFIG} \
     <(echo -e '<ca>') \
     ${KEY_DIR}/ca.crt \
     <(echo -e '</ca>\n<cert>') \
-    ${KEY_DIR}/${1}.crt \
+    ${KEY_DIR}/issued/${1}.crt \
     <(echo -e '</cert>\n<key>') \
-    ${KEY_DIR}/${1}.key \
+    ${KEY_DIR}/private/${1}.key \
     <(echo -e '</key>\n<tls-auth>') \
     ${KEY_DIR}/ta.key \
     <(echo -e '</tls-auth>') \